Try some refactoring
