+++ apps/www/package.json
@@ -53,7 +53,6 @@
     "xss": "1.0.7"
   },
   "devDependencies": {
-    "@foo-software/lighthouse-check": "0.0.90",
     "@types/classnames": "2.2.10",
     "@types/commonmark": "0.27.4",
     "@types/cookie-parser": "1.4.2",
