+++ package.json
@@ -25,6 +25,7 @@
     ]
   },
   "devDependencies": {
+    "@foo-software/lighthouse-check": "0.0.90",
     "@octokit/rest": "16.43.1",
     "@types/fs-extra": "9.0.1",
     "@types/node": "14.0.12",
@@ -39,6 +40,7 @@
     "fs-extra": "9.0.1",
     "husky": "4.2.5",
     "prettier": "2.0.5",
+    "pretty-bytes": "5.3.0",
     "pretty-quick": "2.0.1",
     "ts-node": "8.10.2",
     "typescript": "3.9.5"
