+++ scripts/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "lib": ["esnext"],
+    "lib": ["es2020"],
     "module": "CommonJS",
     "moduleResolution": "node",
     "target": "es5"
