+++ scripts/waitForVercel.ts
@@ -1,4 +1,4 @@
-import Octokit from '@octokit/rest';
+import * as Octokit from '@octokit/rest';
 
 const wait = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
